+++ Facade/UserManager.cs
@@ -6,6 +6,7 @@
     using System;
     using System.Data.Entity;
     using System.Linq;
+    using System.Text;
     using System.Threading.Tasks;
 
     public class UserManager : UserManager<User, int>
@@ -21,11 +22,28 @@
 
         public string ConfirmEmailUrl { get; set; }
 
-        public async Task<UserClaim> AddTempTokenClaimAsync(User user)
+        public override async Task<IdentityResult> AddPasswordAsync(int userId, string password)
         {
-            var claim = Store.AddTempTokenClaim(user);
+            // Add password
+            var result = await base.AddPasswordAsync(userId, password);
+
+            if (result.Succeeded)
+            {
+                // Get user
+                var user = await FindByIdAsync(userId);
+
+                // Remove HasNoPassword claim
+                await Store.RemoveHasNoPasswordClaim(user);
+                await Store.SaveChangesAsync();
+            }
+
+            return result;
+        }
+
+        public async Task AddTempTokenClaimAsync(User user)
+        {
+            await Store.AddTempTokenClaim(user);
             await Store.SaveChangesAsync();
-            return claim;
         }
 
         public async Task<IdentityResult> ChangeEmailAsync(int userId, string email)
@@ -91,8 +109,11 @@
             // Email confirmed
             user.EmailConfirmed = true;
 
+            // Has no password: Determines whether 'Add Password' or 'Change Password' option is available
+            await Store.AddHasNoPasswordClaim(user);
+
             // Temp token: Since this is an external login, create temp token; it's going to be used to retrieve the real token by the client
-            Store.AddTempTokenClaim(user);
+            await Store.AddTempTokenClaim(user);
 
             var result = await base.CreateAsync(user);
 
@@ -100,6 +121,9 @@
             {
                 await Store.AddLoginAsync(user, userLoginInfo);
                 await Store.SaveChangesAsync();
+
+                // Send notification email
+                await SendNewExternalLoginNotificationEmailAsync(user.Id);
             }
 
             return result;
@@ -132,9 +156,8 @@
             if (entity == null)
                 return null;
 
-            // Delete temp token
-            var tempTokenClaim = entity.Claims.Single(claim => claim.ClaimType == "TempToken");
-            await Store.DeleteUserClaimAsync(tempTokenClaim.Id);
+            // Remove temp token
+            await Store.RemoveTempTokenClaim(entity);
             await Store.SaveChangesAsync();
 
             // Return the user
@@ -162,7 +185,7 @@
             user.EmailConfirmed = true;
 
             // Temp token: Since this is an external login, create temp token; it's going to be used to retrieve the real token by the client
-            Store.AddTempTokenClaim(user);
+            await Store.AddTempTokenClaim(user);
 
             var result = await base.AddLoginAsync(user.Id, userLoginInfo);
 
@@ -185,7 +208,7 @@
 
             var confirmEmailUrl = string.Format("{0}?token={1}", ConfirmEmailUrl, encodedToken);
 
-            var sbBody = new System.Text.StringBuilder();
+            var sbBody = new StringBuilder();
             sbBody.AppendLine("    <p>");
             sbBody.AppendLine("        <b>Wealth Economy - Confirm Your Email</b><br />");
             sbBody.AppendLine("        <br />");
@@ -193,6 +216,26 @@
             sbBody.AppendLine("        <br />");
             sbBody.AppendLine("        Please click the following link to confirm your email address<br />");
             sbBody.AppendFormat("        <a href='{0}'>Confirm your email address</a>", confirmEmailUrl);
+            sbBody.AppendLine("    </p>");
+            sbBody.AppendLine("    <p>");
+            sbBody.AppendLine("        Thanks,<br />");
+            sbBody.AppendLine("        forCrowd Foundation");
+            sbBody.AppendLine("    </p>");
+
+            await base.SendEmailAsync(userId, subject, sbBody.ToString());
+        }
+
+        public async Task SendNewExternalLoginNotificationEmailAsync(int userId)
+        {
+            var subject = "New external login";
+            
+            var user = await base.FindByIdAsync(userId);
+
+            var sbBody = new StringBuilder();
+            sbBody.AppendLine("    <p>");
+            sbBody.AppendLine("        <b>Wealth Economy - New External Login</b><br />");
+            sbBody.AppendLine("        <br />");
+            sbBody.AppendFormat("        Email: {0}<br />", user.Email);
             sbBody.AppendLine("    </p>");
             sbBody.AppendLine("    <p>");
             sbBody.AppendLine("        Thanks,<br />");