+++ SolutionItems/Documents/Release Notes.md
@@ -1,5 +1,11 @@
 ﻿### Release Notes
 
+**0.41.1**
+
+* Add password option for new social logins case
+* Send notification email to admin in social login case
+* Update elmah config to use existing smtp settings
+
 **0.41**
 
 * Social logins: Facebook, Google, Microsoft